fix main imports
