+++ main/src/constructions/index.ts
@@ -1,4 +1,5 @@
-export * from "./noop_crdt";
-export * from "./resettable";
-export * from "./semidirect_product";
+export * from "./composite_crdt";
+export * from "./primitive_crdt";
+export * from "./reset_wrapper_crdt";
 export * from "./semidirect_product_rev";
+export * from "./semidirect_product";
