+++ main/src/core/index.ts
@@ -1,5 +1,5 @@
-export * from "./composite_crdt";
+export * from "./causal_broadcast_network";
 export * from "./crdt";
-export * from "./interfaces";
-export * from "./primitive_crdt";
-export { Runtime } from "./runtime";
+export * from "./default_causal_broadcast_network";
+export * from "./runtime";
+export * from "./vector_clock";
