+++ main/src/types/index.ts
@@ -1,7 +1,4 @@
 export * from "./boolean";
-export * from "./core";
-export * from "./helper_crdts";
-export * from "./json";
 export * from "./list";
 export * from "./map";
 export * from "./number";
