+++ main/src/types/list/list_from_map.ts
@@ -1,4 +1,4 @@
-import { CrdtParent } from "../core";
+import { CrdtParent } from "../../core";
 import { CMap } from "../map";
 import { AbstractCListCompositeCrdt } from "./abstract_list";
 import { DenseLocalList } from "./dense_local_list";
