+++ main/src/types/map/mut_map_from_set.ts
@@ -4,7 +4,7 @@ import {
   ElementSerializer,
   PairSerializer,
 } from "../../util";
-import { Crdt } from "../core";
+import { Crdt } from "../../core";
 import { CRegisterEntryMeta } from "../register";
 import { CSet } from "../set";
 import { AbstractCMapCompositeCrdt } from "./abstract_map";
