+++ main/src/util/index.ts
@@ -5,3 +5,4 @@ export * from "./serialization";
 export * from "./weak_value_map";
 export { createRBTree, fillRBTree, Tree as RBTree };
 export * from "./optional";
+export * from "./testing_network";
