+++ package.json
@@ -26,7 +26,7 @@
     "typescript": "^4.3.5"
   },
   "workspaces": [
-    "client",
+    "main",
     "server",
     "demo",
     "benchmarks"
